refactor: extract database and popup to singleton
