--- src/donation.ts
@@ -1,6 +1,0 @@
-export interface Donation {
-	username: string;
-	message: string;
-	amount: number;
-	created: string; // ISO string, sort on date
-}
